исправлен lesson 2 и решен lesson 3!!!

+++ src/lesson3/task1/Loop.kt
@@ -219,7 +219,7 @@ fun sin(x: Double, eps: Double): Double {
         n += 2
     }
     sinx += x1.pow(n + 2) / factorial(n + 2)
-    return sinx.toInt().toDouble()
+    return sinx
 }
 
 /**
@@ -245,7 +245,7 @@ fun cos(x: Double, eps: Double): Double {
         n += 2
     }
     cosx += x1.pow(n + 2) / factorial(n + 2)
-    return cosx.toInt().toDouble()
+    return cosx
 }
 
 /**
